refactoring with moving shared components to common module
