+++ models/multi_fast_pitch.py
@@ -1,4 +1,3 @@
-import copy
 from pathlib import Path
 from typing import Union, Callable, Dict, Any, Optional
 
@@ -6,105 +5,10 @@ import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from torch.nn import Embedding, LayerNorm, MultiheadAttention
+from torch.nn import Embedding
 
-from models.common_layers import LengthRegulator, PositionalEncoding
+from models.common_layers import LengthRegulator, ForwardTransformer, make_token_len_mask
 from utils.text.symbols import phonemes
-
-
-def generate_square_subsequent_mask(sz: int) -> torch.Tensor:
-    mask = torch.triu(torch.ones(sz, sz), 1)
-    mask = mask.masked_fill(mask == 1, float('-inf'))
-    return mask
-
-
-def make_token_len_mask(x: torch.Tensor) -> torch.Tensor:
-    return (x == 0).transpose(0, 1)
-
-
-def make_mel_len_mask(x: torch.Tensor, mel_lens: torch.Tensor) -> torch.Tensor:
-    len_mask = torch.zeros((x.size(0), x.size(1))).bool().to(x.device)
-    for i, mel_len in enumerate(mel_lens):
-        len_mask[i, mel_len:] = True
-    return len_mask
-
-
-class FFTBlock(nn.Module):
-
-    def __init__(self,
-                 d_model: int,
-                 nhead: int,
-                 conv1_kernel: int,
-                 conv2_kernel: int,
-                 d_fft: int,
-                 dropout: float = 0.1):
-        super(FFTBlock, self).__init__()
-        self.self_attn = MultiheadAttention(d_model, nhead, dropout=dropout)
-        self.dropout = nn.Dropout(dropout)
-        self.conv1 = nn.Conv1d(in_channels=d_model, out_channels=d_fft,
-                               kernel_size=conv1_kernel, stride=1, padding=conv1_kernel // 2)
-        self.conv2 = nn.Conv1d(in_channels=d_fft, out_channels=d_model,
-                               kernel_size=conv2_kernel, stride=1, padding=conv2_kernel // 2)
-        self.norm1 = LayerNorm(d_model)
-        self.norm2 = LayerNorm(d_model)
-        self.dropout1 = nn.Dropout(dropout)
-        self.dropout2 = nn.Dropout(dropout)
-        self.activation = torch.nn.ReLU()
-
-    def forward(self,
-                src: torch.Tensor,
-                src_pad_mask: Optional[torch.Tensor] = None) -> torch.Tensor:
-        src2 = self.self_attn(src, src, src,
-                              attn_mask=None,
-                              key_padding_mask=src_pad_mask)[0]
-        src = src + self.dropout1(src2)
-        src = self.norm1(src)
-        src = src.transpose(0, 1).transpose(1, 2)
-        src2 = self.conv1(src)
-        src2 = self.activation(src2)
-        src2 = self.conv2(src2)
-        src = src + self.dropout2(src2)
-        src = src.transpose(1, 2).transpose(0, 1)
-        src = self.norm2(src)
-        return src
-
-
-class ForwardTransformer(torch.nn.Module):
-
-    def __init__(self,
-                 d_model: int,
-                 d_fft: int,
-                 layers: int,
-                 heads: int,
-                 conv1_kernel: int,
-                 conv2_kernel: int,
-                 dropout: float = 0.1,
-                 ) -> None:
-        super().__init__()
-
-        self.d_model = d_model
-        self.pos_encoder = PositionalEncoding(d_model, dropout)
-        encoder_layer = FFTBlock(d_model=d_model,
-                                 nhead=heads,
-                                 d_fft=d_fft,
-                                 conv1_kernel=conv1_kernel,
-                                 conv2_kernel=conv2_kernel,
-                                 dropout=dropout)
-        encoder_norm = LayerNorm(d_model)
-        self.layers = nn.ModuleList([copy.deepcopy(encoder_layer)
-                                     for _ in range(layers)])
-        self.norm = encoder_norm
-
-    def forward(self,
-                x: torch.Tensor,
-                src_pad_mask: Optional[torch.Tensor] = None) -> torch.Tensor:  # shape: [N, T]
-        x = x.transpose(0, 1)  # shape: [T, N]
-        x = self.pos_encoder(x)
-        for layer in self.layers:
-            x = layer(x, src_pad_mask=src_pad_mask)
-        x = self.norm(x)
-        x = x.transpose(0, 1)
-        return x
 
 
 class SeriesPredictor(nn.Module):
